added to exception handling

+++ src/main/java/com/hakancivelek/product_api/dto/ProductResponse.java
@@ -1,8 +1,9 @@
 package com.hakancivelek.product_api.dto;
 
-public record ProductResponse(
-        String id,
-        String name,
-        double price
+import java.math.BigDecimal;
+
+public record ProductResponse(String id,
+                              String name,
+                              BigDecimal price
 ) {
 }
